requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.38.1
python-multipart==0.0.6
pydantic==2.5.0
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.6
aiofiles==23.2.1
google-generativeai==0.3.2
google-cloud-storage==2.10.0
numpy==1.24.3
Pillow==10.1.0